refactoring: moving the otf-tables into their respective packages
