+++ fontbox/src/main/java/org/apache/fontbox/ttf/table/common/RangeRecord.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.fontbox.ttf.gsub;
+package org.apache.fontbox.ttf.table.common;
 
 /**
  * This class models the
