+++ fontbox/src/main/java/org/apache/fontbox/ttf/table/gsub/LigatureSetTable.java
@@ -15,43 +15,42 @@
  * limitations under the License.
  */
 
-package org.apache.fontbox.ttf.gsub;
+package org.apache.fontbox.ttf.table.gsub;
 
 /**
  * This class is a part of the <a href="https://docs.microsoft.com/en-us/typography/opentype/spec/gsub">GSUB — Glyph
  * Substitution Table</a> system of tables in the Open Type Font specs. This is a part of the <a href=
  * "https://docs.microsoft.com/en-us/typography/opentype/spec/gsub#lookuptype-4-ligature-substitution-subtable">LookupType
- * 4: Ligature Substitution Subtable</a>. It specifically models the
- * <a href= "https://docs.microsoft.com/en-us/typography/opentype/spec/gsub#41-ligature-substitution-format-1">Ligature
- * Substitution Format 1</a>.
+ * 4: Ligature Substitution Subtable</a>. It specifically models the <a href=
+ * "https://docs.microsoft.com/en-us/typography/opentype/spec/gsub#41-ligature-substitution-format-1">LigatureSet table:
+ * All ligatures beginning with the same glyph</a>.
  *
  */
-public class LookupTypeLigatureSubstitutionSubstFormat1 extends LookupSubTable
+public class LigatureSetTable
 {
-    private final LigatureSetTable[] ligatureSetTables;
+    private final int ligatureCount;
+    private final LigatureTable[] ligatureTables;
 
-    public LookupTypeLigatureSubstitutionSubstFormat1(int substFormat, CoverageTable coverageTable,
-            LigatureSetTable[] ligatureSetTables)
+    public LigatureSetTable(int ligatureCount, LigatureTable[] ligatureTables)
     {
-        super(substFormat, coverageTable);
-        this.ligatureSetTables = ligatureSetTables;
+        this.ligatureCount = ligatureCount;
+        this.ligatureTables = ligatureTables;
     }
 
-    @Override
-    public int doSubstitution(int gid, int coverageIndex)
+    public int getLigatureCount()
     {
-        throw new UnsupportedOperationException();
+        return ligatureCount;
     }
 
-    public LigatureSetTable[] getLigatureSetTables()
+    public LigatureTable[] getLigatureTables()
     {
-        return ligatureSetTables;
+        return ligatureTables;
     }
 
     @Override
     public String toString()
     {
-        return String.format("%s[substFormat=%d]",
-                LookupTypeLigatureSubstitutionSubstFormat1.class.getSimpleName(), getSubstFormat());
+        return String.format("%s[ligatureCount=%d]", LigatureSetTable.class.getSimpleName(),
+                ligatureCount);
     }
 }
