+++ fontbox/src/main/java/org/apache/fontbox/ttf/table/gsub/package.html
@@ -20,6 +20,7 @@
 
 </head>
 <body>
-This package contains utility classes which extract meanigful data from the highly obfuscated <a href="https://docs.microsoft.com/en-us/typography/opentype/spec/gsub">GSUB table structures</a>.
+This package contains the highly obfuscated <a href="https://docs.microsoft.com/en-us/typography/opentype/spec/gsub">GSUB, aka, Glyph Substitution Table</a> system of tables from the Open Type Font specs. 
+This package makes an honest attempt to closely model these highly obfuscated OTF Tables in the GSUB sub-category, in their raw glory, as close as possible to their original nonsensical forms, without any spite or bias.
 </body>
 </html>
